Split can setup into hondata, all builds fine!
